requirements.txt
fast-api>=0.104.0
uvicorn>=0.24.0
motor>=3.3.1
pymongo>=4.6.0
python-dotenv>=1.0.0
pydantic>=2.4.2
typing>=3.7.4.3
